order controller test class  moved to controller test  package
