+++ server/cart-service/src/test/java/com/fnd/games_store/cart/test/cart_controller_integration_test/OrderController_ProcessOrderTest.java
@@ -1,4 +1,4 @@
-package com.fnd.games_store.cart.test.order_controller_integration_test;
+package com.fnd.games_store.cart.test.cart_controller_integration_test;
 
 
 import com.fnd.games_store.cart.CartApplication;
